Create nodes package
